requirements.txt
musicalbeeps==0.2.9
numpy==1.19.4
Pillow==8.0.1
simpleaudio==1.0.4
